requirements.txt
Pillow
opencv-python
pillow-heif